MEMORY LEAK IN THE SHADER CODE scenes and scene management finished main menu and game scene added.

+++ src/framework/rendering/SpriteSheetFramePicker.cpp
@@ -1,14 +1,13 @@
 #include "SpriteSheetFramePicker.h"
 
 #include "Rectangle.h"
-#include "glm/glm.hpp"
 
 void SpriteSheetFramePicker::PickFrame(const float widthSprites, const  float heightSprites, const  int frameToPick, Rectangle* rectangle)
 {
 	float xStep = (1 / widthSprites);
 	float yStep = (1 / heightSprites);
 	int column = frameToPick % (int)widthSprites;
-	int row = frameToPick / widthSprites;
+	int row = static_cast<int>(frameToPick / widthSprites);
 
 	float xStart = static_cast<float>(column) * xStep;
 	float yStart = static_cast<float>(row) * yStep;
